updated props and get rrquest , and orgnazie code
