+++ src/components/dashboard/teacher/teacherPanels/charts/tables/incidentsByStudentTable.js
@@ -1,6 +1,4 @@
-import React, { useEffect, useState } from 'react';
-import axios from 'axios';
-import { baseUrl } from '../../../../utils/jsonData';
+import React from 'react';
 import { Table, TableContainer, TableHead, TableBody, TableRow, TableCell, Paper, Typography } from '@mui/material';
